Remove code duplication vol 2.
Fix tests in 0.13.
