+++ ci-scripts/sbtTests-0.13.sh
@@ -1,3 +1,3 @@
 set -o xtrace && \
-    sbt ";^^ 0.13.15; test" && \
-    sbt ";^^ 0.13.15; scripted"
+    sbt ";^^ 0.13.16; test" && \
+    sbt ";^^ 0.13.16; scripted"
