+++ hoarder/src/test/sbt_0.13/sbt/inc/MappedTextAnalysisFormatTest.scala
@@ -4,10 +4,7 @@ import java.io.File
 
 import org.scalacheck.Prop
 import org.scalacheck.Properties
-import sbt.inc.hoarder.DefaultFormat
-import sbt.inc.hoarder.MappableFormat
-import sbt.internal.inc._
-import sbt.internal.inc.hoarder.DefaultFormat
+import sbt.inc.hoarder._
 
 import scala.util.Random
 import scala.util.Try
